refactor!: move operational baggage manager from sdk to api
