+++ api/lib/opentelemetry/baggage.rb
@@ -4,10 +4,12 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+require 'opentelemetry/baggage/propagation'
 require 'opentelemetry/baggage/builder'
 require 'opentelemetry/baggage/entry'
 require 'opentelemetry/baggage/manager'
-require 'opentelemetry/baggage/propagation'
+require 'opentelemetry/baggage/noop_builder'
+require 'opentelemetry/baggage/noop_manager'
 
 module OpenTelemetry
   # The Baggage module provides functionality to record and propagate
