+++ api/lib/opentelemetry/baggage/builder.rb
@@ -6,13 +6,33 @@
 
 module OpenTelemetry
   module Baggage
-    # No op implementation of Baggage::Builder
+    # SDK implementation of Baggage::Builder
     class Builder
-      def set_entry(key, value, metadata: nil); end
+      attr_reader :entries
 
-      def remove_entry(key); end
+      def initialize(entries)
+        @entries = entries
+      end
 
-      def clear; end
+      # Set key-value in the to-be-created baggage
+      #
+      # @param [String] key The key to store this value under
+      # @param [String] value String value to be stored under key
+      def set_entry(key, value, metadata: nil)
+        @entries[key] = OpenTelemetry::Baggage::Entry.new(value, metadata)
+      end
+
+      # Removes key from the to-be-created baggage
+      #
+      # @param [String] key The key to remove
+      def remove_entry(key)
+        @entries.delete(key)
+      end
+
+      # Clears all baggage from the to-be-created baggage
+      def clear
+        @entries.clear
+      end
     end
   end
 end
