+++ propagator/ottrace/test/opentelemetry/propagator/ottrace/text_map_injector_test.rb
@@ -119,7 +119,7 @@ describe OpenTelemetry::Propagator::OTTrace::TextMapInjector do
 
     describe 'baggage handling' do
       before do
-        OpenTelemetry.baggage = OpenTelemetry::SDK::Baggage::Manager.new
+        OpenTelemetry.baggage = OpenTelemetry::Baggage::Manager.new
       end
 
       after do
