+++ sdk/lib/opentelemetry/sdk.rb
@@ -73,7 +73,6 @@ module OpenTelemetry
 end
 
 require 'opentelemetry/sdk/configurator'
-require 'opentelemetry/sdk/baggage'
 require 'opentelemetry/sdk/internal'
 require 'opentelemetry/sdk/instrumentation_library'
 require 'opentelemetry/sdk/resources'
