--- sdk/lib/opentelemetry/sdk/baggage.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-# Copyright The OpenTelemetry Authors
-#
-# SPDX-License-Identifier: Apache-2.0
-
-require 'opentelemetry/sdk/baggage/builder'
-require 'opentelemetry/sdk/baggage/manager'
-
-module OpenTelemetry
-  module SDK
-    # Contains operational implementations of the Baggage::Manager
-    module Baggage
-    end
-  end
-end
